Un-mopafy Event trait so it can be parameterized by Error
This makes storing and retrieving the Events quite a bit more
complicated because of the need for a wrapper type. Though it might
be possible to make it better.  The good news is, that the majority
use case of only accessing Event data inside the handler for that
specific event is still easy with the &self parameter.