Change effect store to vector rather than AnyMap
It seems like a way more niche use-case to be able to retrieve the
Effects by type, so it's not necessary to have them be stored in a map
structure.  They only need to be in a map in re-frame because it's the
key that actually identifies what Effect to call.  But with types and
a method on the type, we can just call the method on the Effect
struct without having any idea what it is.